\name{cytof2-package}
\alias{cytof2-package}
\alias{cytof2}
\docType{package}
\title{
\packageTitle{cytof2}
}
\description{
\packageDescription{cytof2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cytof2}
\packageIndices{cytof2}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{cytof2}

Maintainer: \packageMaintainer{cytof2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
