% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat.R
\name{invgamma_ab}
\alias{invgamma_ab}
\title{Get the parameter values for inv-gamma distribution given mean and sd.
Parameterization: mu = b / (a-1)}
\usage{
invgamma_ab(mu, sig)
}
\description{
Get the parameter values for inv-gamma distribution given mean and sd.
Parameterization: mu = b / (a-1)
}
