% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta.R
\name{p_missing}
\alias{p_missing}
\title{Compute probability of missing across a grid of values (y)}
\usage{
p_missing(o, i, j, y)
}
\arguments{
\item{o}{A list containing the parameters from one iteration of MCMC}

\item{i}{Index for sample}

\item{j}{Index for marker}

\item{y}{A grid of values to compute the probabilities}
}
\description{
Compute probability of missing across a grid of values (y)
}
