% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta_new.R
\name{p_missing_new}
\alias{p_missing_new}
\title{Compute probability of missing across a grid of values (y)}
\usage{
p_missing_new(o, i, j, y, c0)
}
\arguments{
\item{o}{A list containing the parameters from one iteration of MCMC}

\item{i}{Index for sample}

\item{j}{Index for marker}

\item{y}{A grid of values to compute the probabilities}

\item{c0}{Basically a spline point}
}
\description{
Compute probability of missing across a grid of values (y)
}
