% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_alloc.R
\name{pairwise_alloc}
\alias{pairwise_alloc}
\title{Returns a pairwise allocation matrix of a feature allocation (binary) matrix Z
using SALSO by David B. Dahl.}
\usage{
pairwise_alloc(Z)
}
\description{
Returns a pairwise allocation matrix of a feature allocation (binary) matrix Z
using SALSO by David B. Dahl.
}
