% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta_new.R
\name{plot_prob_missing_new}
\alias{plot_prob_missing_new}
\title{Plot the posterior distribution of p_inj across grid of values (y_grid)}
\usage{
plot_prob_missing_new(mcmc, i, j, c0, q = c(0.025, 0.975), y_grid = seq(-12,
  12, len = 50), plot_line = TRUE, plot_area = TRUE,
  addToExistingPlot = FALSE, col.line = "blue", col.area = rgb(0, 0, 1,
  0.1), ...)
}
\arguments{
\item{mcmc}{The output from cytof_fit}
}
\description{
Plot the posterior distribution of p_inj across grid of values (y_grid)
}
