% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat.R
\name{simdat}
\alias{simdat}
\title{Generate simulation data}
\usage{
simdat(I, N, J, K, W, Z = genZ(J, K), gams_0 = matrix(1/rgamma(I * J, 3,
  0.1), nrow = I), sig2 = matrix(1/rgamma(I * J, 3, 1), nrow = I),
  psi_0 = -1, psi_1 = 1, tau2_0 = 2, tau2_1 = 1, thresh = -5)
}
\description{
Generate simulation data
}
