% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta.R
\name{get_beta_new}
\alias{get_beta_new}
\title{Compute betas from y and p_target (new version)}
\usage{
get_beta_new(y = c(-3, -2, -1), p_target = c(), plotting = FALSE)
}
\description{
Compute betas from y and p_target (new version)
}
